jwtapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
email-validator
passlib[bcrypt]
pyjwt
python-multipart